backend/requirements.txt
# Web框架
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# AI API客户端
openai==2.6.0

# HTTP客户端
httpx==0.28.1
requests==2.32.3

# 配置管理
pydantic==2.10.6
pydantic-settings==2.7.1

# 重试机制
tenacity==9.0.0

# 测试
pytest==8.3.5
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# 代码质量
black==24.10.0
mypy==1.14.1
ruff==0.9.2

# 环境变量
python-dotenv==1.0.1

# 阿里云OSS
oss2==2.18.4

# 数据库
sqlalchemy==2.0.36
alembic==1.14.0

# 密码哈希
passlib[bcrypt]==1.7.4

# JWT认证
python-jose[cryptography]==3.3.0

# Google GenAI SDK
google-genai==1.46.0
